# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "description": "Pet Shop API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "email": "contact@asyncdeveloper.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Pet Shop API"
        }
    ],
    "paths": {
...
        "/api/v1/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthourized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "login a user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthourized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get current user",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/users/edit": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "update current user",
                "operationId": "updateUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthourized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "forgot password",
                "operationId": "forgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "e1adfed7194b11c9641415b0c937b8b4b8f11a3794907f8ab75d108c94941e79"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthourized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/reset-password-token": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "reset password token",
                "operationId": "resetPasswordToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthourized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/users/order": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all user orders",
                "operationId": "getUserOrder",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "path",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "logout current user",
                "operationId": "logoutUser",
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/delete": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "delete current user",
                "operationId": "deleteUser",
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthourized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateUserRequest": {
                "title": "CreateUserRequest",
                "description": "register a new user",
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "avatar",
                    "address",
                    "phone_number",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "Oluwaseyi"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Adeogun"
                    },
                    "email": {
                        "type": "string",
                        "example": "contact@asyncdeveloper.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://placehold.co/600x400/png"
                    },
                    "address": {
                        "type": "string",
                        "example": "Lagos,Nigeria"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08165551288"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "is_marketing": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ForgotPasswordRequest": {
                "title": "ForgotPasswordRequest",
                "description": "forgot password",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "contact@asyncdeveloper.com"
                    }
                },
                "type": "object"
            },
            "LoginUserRequest": {
                "title": "LoginUserRequest",
                "description": "login a user",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "contact@asyncdeveloper.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
...
            "PasswordResetRequest": {
                "title": "PasswordResetRequest",
                "description": "password reset",
                "required": [
                    "email",
                    "token",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "e1adfed7194b11c9641415b0c937b8b4b8f11a3794907f8ab75d108c94941e79"
                    },
                    "email": {
                        "type": "string",
                        "example": "contact@asyncdeveloper.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
            "Order": {
                "title": "Order",
                "description": "Order model",
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}